Move execute_wait into parsl.utils and remove parsl.channels remnant
